==> python/0_download_dataverse.ipynb <==
from pyDataverse.api import NativeApi, DataAccessApi
from constants import DATA_FOLDER
import os

# Create a new data directory if it does not exist
if not os.path.exists(DATA_FOLDER):
    os.makedirs(DATA_FOLDER)

# Initialize api access
base_url = "https://dataverse.harvard.edu/"

api = NativeApi(base_url)
data_api = DataAccessApi(base_url)

# Download from harvard dataverse the sitc trade data, basic classifications, and product space
# For robustness check, consider switching to the HS classification DOI_hs = "doi:10.7910/DVN/T4CHWJ" and download corresponding dataset
DOI_sitc = "doi:10.7910/DVN/H8SFD2"
DOI_ps = "doi:10.7910/DVN/FCDZBN"
dataset_sitc = api.get_dataset_version(DOI_sitc, version="4.0")
dataset_ps = api.get_dataset_version(DOI_ps, version="1.1")

files_sitc = dataset_sitc.json()["data"]["files"]
files_ps = dataset_ps.json()["data"]["files"]

for file in files_sitc:
    filename = file["dataFile"]["filename"]
    file_id = file["dataFile"]["id"]
    if filename.startswith("country_sitcproduct4digit"):
        response = data_api.get_datafile(file_id, data_format="original")
        with open(f'{DATA_FOLDER}/{file["dataFile"]["originalFileName"]}', "wb") as f:
            f.write(response.content)

for file in files_ps:
    filename = file["dataFile"]["filename"]
    file_id = file["dataFile"]["id"]
    if filename.startswith("sitc2_proximities"):
        response = data_api.get_datafile(file_id, data_format="original")
        with open(f'{DATA_FOLDER}/{file["dataFile"]["originalFileName"]}', "wb") as f:
            f.write(response.content)
